docker_workloads/image_rec/resnet/requirements.txt
Flask==2.2.3
Pillow==9.2.0
Pillow==9.5.0
Requests==2.28.2
torch==1.12.1+cu113
torchvision==0.13.1+cu113
